+++ pkg/timeDate/R/holiday-LONDON.R
@@ -20,36 +20,33 @@
 #  holidayLONDON             Returns holidays for British Bank Holidays
 ################################################################################
 
-# ---------------------------------------------------------------------------- #
-# Roxygen Tags
-#' @export
-# ---------------------------------------------------------------------------- #
 holidayLONDON <- function (year = getRmetricsOptions("currentYear")) {
     ## function implemented by Menon Murali;
     ## updated, corrected and refactored by GNB
 
+    ## ignoring year <= 1833 as there wereno official holidays then
+    y1834to1870 <- year[year >= 1834 & year <= 1870]
+    y1871toInf <- year[year >= 1871]
+
     holidays <- list() # 2023-05-12 GNB was: NULL
-    for (y in year) {
-        if (y >= 1834 & y <= 1870) {
-            # 1 May, 1 November, Good Friday and Christmas are the
-            # only England Bank Holidays
-            dts <- c(paste0(y, "-05-01"), paste0(y, "-11-01"))
-            holidays <- list(holidays, dts, GoodFriday(y, ""),
-                          ChristmasDay(y, ""))
-        }
-        if (y >= 1871) {
-            holidays <- list(holidays,
-                          ## as.character(GoodFriday(y)),
-                          ## as.character(EasterMonday(y)),
-                          ## as.character(GBSpringBankHoliday(y)),
-                          ## as.character(GBSummerBankHoliday(y))
 
-                          GoodFriday(y, ""),
-                          EasterMonday(y, ""),
-                          GBSpringBankHoliday(y, ""),
-                          GBSummerBankHoliday(y, "")
-                          )
+    if(length(y1834to1870)) {
+        ## 1 May, 1 November, Good Friday and Christmas are the
+        ## only England Bank Holidays
+        holidays <- list(holidays,
+                         paste0(y1834to1870, "-05-01"),
+                         paste0(y1834to1870, "-11-01"),
+                         GoodFriday(y1834to1870, ""),
+                         ChristmasDay(y1834to1870, ""))
+    }
 
+    if(length(y1871toInf)) {
+        holidays <- list(holidays,
+                         GoodFriday(y1871toInf, ""),
+                         EasterMonday(y1871toInf, ""),
+                         GBSpringBankHoliday(y1871toInf, ""),
+                         GBSummerBankHoliday(y1871toInf, "")
+                         )
             # Not entirely sure when Mon/Tue began to be given as
             # holiday when Christmas/Boxing Day fell on
             # Saturday/Sun. I'm assuming this was after the 1971
@@ -71,93 +68,64 @@ holidayLONDON <- function (year = getRmetricsOptions("currentYear")) {
             ##
             ## GNB: the logic can be simplified by adding Christmas and Boxing days and the
             ##      Mon/Tue where necessary. Sat/Sun are omitted at the end anyway.
-            holidays <- list(holidays, as.character(
-                if (y < 1970) {
-                    c(ChristmasDay(y), BoxingDay(y))
-                } else {
-                    posix1 <- as.POSIXlt(ChristmasDay(y))
-                    switch(as.character(posix1$wday),
-                           ## (posix1$wday == 0)  # Christmas on Sunday
-                           "0" =  ChristmasDay(y) + (1 : 2) * 86400,
-                           
-                           ## (posix1$wday == 6) # Christmas on Saturday
-                           "6" = ChristmasDay(y) + (2 : 3) * 86400,
-                    
-                           ## posix1$wday == 5) # Christmas on Friday
-                           ##     the next Monday is a holiday
-                           "5" = ChristmasDay(y) + c(0, 3) * 86400,
-                           
-                           ## default 1,2,3,4
-                           c(ChristmasDay(y), BoxingDay(y))
-                           )
-            }))
-
-            if (y >= 1974) {
-                ## New Year's Day: if it falls on Sat/Sun, then is
-                ## moved to following Monday
-                posix1 <- as.POSIXlt(NewYearsDay(y))
-                holidays <- list(holidays, 
-                    if (posix1$wday == 0 | posix1$wday == 6) {
-                        format..sdate(.on.or.after(y, 1, 1, 1))
-                    } else {
-                        NewYearsDay(y, "")
-                    })
-            }
-            
-            if (y >= 1978) {
-                ## Early May Bank Holiday
-                holidays <- list(holidays, GBEarlyMayBankHoliday(y, ""))
-                
-                ## special one-off bank holidays
-                if (y == 1981) {
-                    ## Royal wedding was a public holiday
-                    dts <- paste0(y, "-07-29")
-                    holidays <- list(holidays, dts)
-                } else if (y == 1999) {
-                    ## UK millenum day
-                    dts <- format(GBMilleniumDay(), format = "%Y-%m-%d")
-                    holidays <- list(holidays, dts)
-                } else if (y == 2002) {
-                    # Last Monday in May holiday moved to June 3, and
-                    # Queen's Jubilee on June 4
-                    dts <- list(# paste0(y, "-06-03"),
-                             paste0(y, "-06-04"))
-                    holidays <- list(holidays, dts)                
-                } else if (y == 2011) {
-                    ## Royal wedding declared a public holiday
-                    dts <- paste0(y, "-04-29")
-                    holidays <- list(holidays, dts)
-                } else if (y == 2012) {
-                    # Last Monday in May holiday moved to June 4, and
-                    # Queen's Diamond Jubilee on June 5
-                    dts <- list(# paste0(y, "-06-04"),
-                             paste0(y, "-06-05"))
-                    holidays <- list(holidays, dts)
-                } else if (y == 2022) {
-                    ## Last Monday in May (i.e., Spring Bank Holiday) holiday moved to June 2,
-                    ## Unique Bank holidays:
-                    ##     Queen's Diamond Jubilee.
-                    ##     State Funeral of Queen Elizabeth II
-                    dts <- list(# paste0(y, "-06-02"), # Thursday, Spring bank holiday
-                             paste0(y, "-06-03"), # Friday, Platinum Jubilee bank holiday
-                             paste0(y, "-09-19")  # Bank Holiday for the State Funeral of
-                                                  # Queen Elizabeth II
+        ## borderline 1969 needs checking
+        y1871to1969 <- y1871toInf[flag1969 <- y1871toInf <= 1969]
+        y1970toInf  <- y1871toInf[!flag1969]
+        
+        if(length(y1871to1969)) {
+            holidays <- list(holidays,
+                             ChristmasDay(y1871to1969, ""),
+                             BoxingDay(y1871to1969, "")
                              )
-                    holidays <- list(holidays, dts)
-                } else if (y == 2023) {
-                    ## Bank holiday for the coronation of King Charles III
-                    dts <- paste0(y, "-05-08")
-                    holidays <- list(holidays, dts)
-                }
-            }
         }
+
+        if(length(y1970toInf)) {
+            xmas <- ChristmasDay(y1970toInf)
+            posix1 <- as.POSIXlt(xmas)
+            wd <- as.character(posix1$wday)
+            f <- function(x, y) {
+                switch(x,
+                       ## (posix1$wday == 0)  # Christmas on Sunday
+                       "0" =  as.character(ChristmasDay(y) + (1 : 2) * 86400),
+                       
+                       ## (posix1$wday == 6) # Christmas on Saturday
+                       "6" = as.character(ChristmasDay(y) + (2 : 3) * 86400),
+                                   
+                       ## posix1$wday == 5) # Christmas on Friday
+                       ##     the next Monday is a holiday
+                       "5" = as.character(ChristmasDay(y) + c(0, 3) * 86400),
+                       
+                       ## default 1,2,3,4
+                       list(ChristmasDay(y, ""), BoxingDay(y, ""))
+                       )
+            }
+            holidays <- list(holidays,
+                             .mapply(f, list(wd, y1970toInf), NULL)
+                             )
+        }
+
+        y1974toInf  <- y1970toInf[y1970toInf >= 1974]
+        if(length(y1974toInf)) {
+            ## New Year's Day: if it falls on Sat/Sun, then is
+            ## moved to following Monday
+            posix1 <- as.POSIXlt(NewYearsDay(y1974toInf))
+            weekend <- posix1$wday == 0 | posix1$wday == 6
+        
+            holidays <- list(holidays,
+                             format..sdate(.on.or.after(y1974toInf[weekend], 1, 1, 1)),
+                             NewYearsDay(y1974toInf[!weekend], "")
+                             )
+        }
+
+        ## Early May Bank Holiday year >= 1978
+        holidays <- list(holidays,
+                         GBEarlyMayBankHoliday(y1974toInf[y1974toInf >= 1978], ""),
+                         GBOneoffBankHoliday(y1974toInf, "")
+                         )
     }
 
-#browser()    
-    holidays <- unlist(holidays)
-    holidays <- sort(holidays)
+    holidays <- sort(unlist(holidays))
     ans <- timeDate(holidays, zone = "London", FinCenter = "Europe/London")
     posix1 <- as.POSIXlt(ans, tz = "GMT")
     ans[!(posix1$wday == 0 | posix1$wday == 6)] # Remove any Saturdays/Sundays
 }
-