+++ pkg/timeDate/R/holiday-Listing.R
@@ -77,7 +77,7 @@ listHolidays <-
         "GBEarlyMayBankHoliday", # was: "GBMayDay",
         "GBSpringBankHoliday",   # was: "GBBankHoliday"
         "GBSummerBankHoliday",
-        "GBMilleniumDay",
+        "GBOneoffBankHoliday", # new 2023-12-08
         "DEAscension",
         "DECorpusChristi",
         "DEGermanUnity",